Convert VehicleMission to an interface
This is not a perfect commit as VehicleMission logic is spread around
the simualtion and is not well isolated. This commit focuses on
isolating the "existing" logic. The net effect is that the new
VehicleMission interface has too many methods and exposes too mich of
the internal logic.
However this can be reworked on subsequent changes to make
VehicleMission a good contract interface.

Changes:
- Rename VehicleMission to AbstractVehcileMission
- Create VehicleMission interface
- Create MissiotnUtil to hold the helper methods not specific to
  individual missions.

Part of #625