Optimise the MonitorWindow to reduce the processing load.
Th primary goal of this change is to allow the table models ot be reused when the Settlement selected is changed.
This replaces the previous approach of creating a new tab eveyr time which generates a large number of event listeners
attached to Units & UnitManager.
Secondly, to reduce the numbenr of table redrawn events that are being triggered.

Changes:
- MonitorModel has a setSelectedSettlement method that has to be implemented by subclasses.
- UnitTableModel has a single resetUnits mehtods that can reset all Units modeled in one operation and one table event.
- UnitTableModel holds units as a List which simplifies the indexing and lookup methods.
- Unit listneers are correctly protected with synchronised keyword
- MonitorModel impelemntors reset the cotnents according to the selected Settlement
- Event & Mission models do not support selected Settlement so throw exceptino if called
- MonitorWindow handles the tab selection by using the existing tab without creating a new tab
- Fix code smells

Part of #719