Introduce the concept of TaskJob
The new TaskJob interface represents an activity that has to be performanced via a new Task. It has an associated score property.
The TaskCache now collects holds a collection of TaskJobs instead of a direct connection to the TaskMeta instance.
The purpose of this change is to support more flexibility on how the actual Task instance is created, e.g. extra parameters passed.

Changes:
- New TaskJob interface created
- New BasicTaskJob class created that delegates Task creation to a MetaTask object; this replicates the current approach.
- Person & Robot TaskManager are updated for the default idle task to use sharable TaskJob instances
- Baseclass MetaTask exposes new method to return a list of TaskJob instead of just 1 single probability
- TaskCache is built from TaskJobs, including adding a complete list
Part of #740