Fix problem with new Saving Unit TestSaving
This test worked on it's own but fails when run as part of the suite due to the Simulation object being shared and not truely stateless.
There is a big problem os classes depending upon Simulation.instance method. These classes have the potential to reference the wrong sub-objects during the UnitTest because they get recreated for tests. The use of a instance pattern is dangerous.

We should llok to remvoe the static references if possible or explcitly initialise them.

Changes:
- Initialise Job explicitly from Simulation start up
- UnitManager provide against NPE when started via UnitTest - needs fixing