Tidy up MetaTask code to remove unclear code
MetaTask class according to the definition supports the creation of Tasks without a context. This is the old approach of the MetaTask being a factory for the Task class.
In the new approach only a subset of the MetaTask follwo the factory pattern, namely those that create personalised Tasks.
This is important for the CommanderWindow as that should only allow the factory style metaTasks to be used.
To remove this confussion a new FactoryMetaTask class is now provided that represents the Factory Style metatask.

Changes:
- New FactoryMetaTask class
- Remove old applyPersonModifier code as replaced by standard getPersonModifier
- Add new getBuildingModifier method to check for over crowding and relationships constraints
- Updated MetaTasks to use new modifier methods
- Migrated single use MetaTasks to the FactoryMetaTask class

Part of #740