Create UnitTest for MetaTaskUtil lookups
MetaTaskUtil provides a lookup from a Task to the associated MetatTask. There is no expliclt binding between the two so a class name lookup is used. This is fairly straight forward but of course is fragile to class changes.
This commit adds a UnitTest to check that the commonly used Tasks do work with the lookup method. This will trap any future deviation.

Changes:
- MetaTaskUtilTest tests various Task scenarios. Including the EVA & Garage tasks that map to a single MetaTaskUtil
- RepairMalfunctionMeta, LoadVehicleMeta & UnloadVehicleMeta correct the inaccurate name to be  EVA/Inside agnostic.
- Simulation has more manager classes added to the test setup to support Missions in UnitTests. It's debatable whether the testRun method is any better than the full initiailise now.
- AbstractMarsSimUnitTest supports the creation of a basic Vehicle Mission for use in Tests
- LoadVehicleEVA traps attempts to load a completed mission
- Teach tidy up to use Worker whenever possible
- Fix repating blank entries in the Activity log. This reduce the concurrency problem in the activity Console command as Robots do not continuously update the activity log when charging.
- TendGreenhouse does not record each crop in the Activity log

Part of #740, #749