EquipmentInventory minimises reach for AmountResources
EquipmenetInventory was reaching intot he individual contained Equipment for resource & items. This is increasing the change of concurenncy issues asthe Equipment set is a fluid attribute.
Secondly resources held within the Equipment are not directly accessable to their own; most Equipment transfer their contents to their parent when being transferred so the deep reach is largely redundant.

Relates to #736 #738