Malfunctions creates extra demand for Tasks
Repair tasks can have a Worker demand of more than 1 as there can be multiple repairer slots available.
The SettlementTask now supports a demand property such that a singel task can be resused multiple times, upto the demand value.
The default demand is 1 but for repair tasks it is set to the number of available slots.

Changes:
- SettlementTask maintains a demand value and reduces it whenever a Task is used
- The SettlementTaskManager only removes the shared backlog Tasks once the demand reaches zero.
- Repair tasks are created for ALL active malfunctions; not just the most serious. They also set the demand to the available slot count.
- The SettlementTaskManager has a flag to trigger the tasks to be refreshed. These get flagged for refresh every time pulse but the list is only recreated when a Worker requested the latest backlog tasks. The net effect is that the available backlog Tasks remain visible.
- backlog console command shows the demand for each task.
Part of #141