Add concept of future scheduled events/JUnitTests
A Futuer Scheduled Event is an event that is scheduled to trigger at a specific time in the future. These are managed by a ```ScheduledEventManager```.
A settlement has a dedicated for local events. When an event arrives is it passed to a handler class for execution. The handler can request that the event is rescheduled to repeat in the future.
The goal of this change is to remove the polling of time values during time passing methods.
The Shift logic has been converted to use future events now for Leave, Shift Rotation and Shift on/Off Duty.

Changes:
- New classes ScheduledEventHandler, ScheduledEvent & ScheduledEventHandler.
- ScheduledEventManager maintains an order list of ScheduledEvents according to the event time.
- Settlement creates it's own ScheduledEventManager and includes it in the timePassing method.
- ShiftManager no longer has a timepassing method as it is now event driven. Constructor creates an event for the Shift Rotation logic.
- Shift class acts as a handler for it's own changing of Shift duty.
- ShiftSlot acts as a handler when the owning Person comes off leave. The on leave methof now takes a variable duration for the leave.
- ShiftManager maxOnLeave attribute is renames to maxRotation to better reflect it's purpose.
- Updated the ShiftManagerTest to use ClockPulse to advance the time instead of explcitly polling.
- AbstractMarsSimUnitTest provides methods to create ClockPulses whilst keeping the main mars clock in synch.
- Added a new 'future' Console command for Settlements to show any future events.

Part of #761