Single ClockListner drives the UI windows
This is a big change to how time is propagated in the UI.
The MainWindow class is now the only ClockListener and it propagates the clock changes to the ToolWindow via the update method.
Any panel that was an standalone ClockListners is now updated by their parent ToolWindow.
This means all windows will be synchronised.

Changes:
- ToolWindow takes a ClockPulse as argument in the update method.
- MainWindow passes the clock pulse onto the ToolWindows.
- TimeWindow updates the displayed from the update method instead of a ClockListener
- MapPanel is updatd via an update method from the parent window.
- MissionWindow & NavigatorWindow update the MapPanel when they are notified of a change of clock time.
- MapPanel is not dynamically updated when used in the Create Mission wizard. This is fine because the map is filtered in this situation.
- ResupplyWindow updates the subpanels when the clock changes.
- SettlementWindow is responible for updating the SettlementTransparentPanel as time changes.
- Remaining ToolWindow that override update method have the signature updated

close #800