Avoid updating job during the change
04.09.2023

- change: fine-tune checkJobReassignment() in TabPanelCareer

Note: this address the situation when a player is still in the
process of choosing yes/no/cancel in the JOptionPane regarding the
job change. TabPanelCareer should not compare if the item being chosen
in the job combobox is equal to the one current job type at that
moment.

Part of https://github.com/mars-sim/mars-sim/issues/281