Consolidate UI opening details windows
Create a new Entity interface that represents the top levle objects that the simulation controls.
This entity is used to centralise the opening of the appropriate UI Window to show the details of
the entity. This oculd be a Mission or Unit Window.
The objective is to remove the coupling between the various UI Windows.

Changes:
- Create Entity interface and implement is Unit, Mission & Transportable
- MainDesktopPane replaces the openUnitWindow with a generic showDetails method
- The new showDetails method acts as a factory to select the appropriate UI window to display the selection.
- Update all calls to the old openUnitWindow with the replacement
- Update direct calls to the MissionWindow with the generic Desktop method
- Move the open details window logic into the MonitorWindow class

Part of #827