Fix infinite walk problem
The WalkSettlementInterior Task can go into an infinite loop if there is a very small time period left.
When there is no waling; the time remaining return never changes so the Taskmanager keeps calling
the same Task.

The fix is to calculate the potential covered meters as one calc to remove the rounding to zero issue.