Fix problem with new Settlement after  reload
The problem relates to the MineralMapConfig being serialized in the save which is incorrect.
Config class do not get saved with the simulation. The root cause is the mineral list is transient in the config.
Hence when it is restored the moneral list is null and cresting a new Settlement fails.

The fix is to stop the MineralMapConfig being saved in the sim file. This means the config is correctly reloaded from XML.

Note the RandomMineralMap needs cleaning up.

fix #1005