Improve the MissionTableModel
When a Mission is deleted; the table modle event is not fire immediately but in a background Swing threaad.
This raised the prospect of a race condition.
Change so the event is fired when the cache list is updated to keep in synch.

Also when adding new mission, correct the new row event to cover the correct row range.

Close #1119