ActivitySpots are allocatable to a Worker
This changes how ActivitySpots are allocated. They are now created as part of a Function constructor.
An ActivitySpot now remembers which Worker it is assigned to as it is explcitly allocated to the Worker.
Once allocated, the Worker receives a dedicated AllocatedActivitySpot which allows them to release it once completed.
Only position that have activity spots created can be allocated; this is another different from the previous implementation which was a loose binding.
This has highlighted the complexity in using 2 different coordinate frames (Settlement & local Building); Activity Spots currently are in the Building coordinate frame.

Changes:
- ActivitySpot remembers the Worker assigned; this can later be released by the AllocatedActivitySpot object.
- Function is responsable for creating the required Activity Spots.
- Worker no longer remembers the Function but just the AllocatedActivtySpot.
- A Worker releases any old ActivitySpot when a new one is assigned.
- Arrival and claiming of an ActivitySpot has been changed in many places.
- Corrected a problem with allocating an ActivitySpot for Dining; old approach could assign the wrong Function.
- Airlock has been converted to use ActivitySpot for the main airlock location.
- Airlock methods are now based on Person instead of identifier; internally identifer is still heavily used :-(
- Task walk to bed corrected as previous it could use the wrong coordinate frame to find the Activity Spot.
- At the start of a Walk task, any activity spot is released.
- Changed LocalPosition equals method to have a 1 centimeter tolerance. This is a precision problem with using double and converting between the 2 coordinate frames.
- New Function unit test to check ActivitySpot logic.

Part of #1143