Rework Settlement Map layers
This aims to rework the map layers in the settlement to clean up the code and make it optimal.
The map displays the "items" and then the LabelMapLayer cycle back over the units to render the labels.
This means a double scan of all the itmes being displayed. The aim is to display labels as part of the item rendering.

Changes:
- Create ColorChoice which holds a 2 colour pairing used for rendering text and highlights. Removes the need of r2 variables being passed.
- AbstractMapLayer is new and provides reusable label drawing methods including multi-line for long labels.
- Font and ColorChoice is used as parameters for label drawing
- New WorkerMapLayer renders Worker objects according to user selection. Has a callback for ColorChoice to use.
- Selected Worker is always shown with labels.
- PersonMapLayer & RobotMapLayer use WorkerMapLayer to greatly simplify their implementation
- "unit" layers now draw labels as part of the main layer drawing instead of a seperate layer
- Correct problem with labels being auto-enabled when a Robot/Person is selected. Now respects users choice.
- StructureMapLayer & VehicleMapLayer use LocalBoundingObject for placement instead of 6 seperate parameters.
- Numerous code smells corrections
- Delete LabelMapLayer

Helps to support future change for #1166