Fix Building Category
Issues #1339 highlighted soem Buildings are not gettign the correct category.
The problem lies in the LifeSupprot & RoboticStatino Functions have a Living category assigned. These are always present and hence override any other Functions.
The correction is to allow a FunctionType to have no category, liek the 2 above. Then these are excluded from the Category logic.

Changes:
- Remove category from Life Supprot & RoboticStation
- BuildingConfig can handle a Functiontype not having a category
- Create a new UnitTest to check category of some common buildings