Automatically add Manufacturing processes to the queue
This based on the value to teh Settlement of the Process outputs. The queuable processes are filtered for resoruces available and then valued.
Those above a threshold controlled by a Preference are taken forward as candidates. The candidates are sorted and teh top N are added tothe queue;
where N is defined by another Preference setting. This Preference can be use to control automatic creation.
Lastly there is an overarching Preference to control the size of the queue.

Changes:
- Create new Parameter category called ManufacturingParameters that controls the Preferences for Manufacturing.
- Added new entry in message.properties
- ManufacturingManager automatically adds new ManufactureProcesses to the queue if they meet the qualifying criteria.
- ProcessItemFactory load the id of ProcessItems when the type is Bin or Equipment
- TabPanelPreferences supports the new ManufacturingParameters
- Remove unused methods from ManufactureGood and ManufactureUtil
- Update the UnitTests

Part of #1236