Make creation of SimulationConfig explicit
Previously the SimulationConfig.instance method would create a new instance on demand if one was not created. This is a problem beause this method is called multiple times and hence it means the contruction can not be controlled.
This change ensure a new config instance is created in a determinsitic fashion under full control of the code.
The consequence is that the instance method will throw an expcetion if an config has not been loaded.
Objective of this chnage is to stop using SimualtionConmfig.instance inside class initialiers.

Changes:
- SimulationConfig instance method changed to only return preivously create instance
- SimualtionConfig loadConfig will create a shared instance if needed and load the XML files
- Update UnitTests to use the SimulationConfig.loadConfig approach

Part of #1555