Fix anomalously high demand of resources on high population settlement
2025-07-22

Note: The population should only minimally impact the demand score and
NEVER be proportional to it. If not, a 1000 people settlement
would have a 100 times more demand on oxygen than one with a population
of 10. This will potenially throw the production of oxygen off-balance
as observed in the default scenario. The quantity of a secondary
resource ties to the scale of production and is NOT linearly
proportional to the demand score of a resource

- new: create popFactor to lower the spike of demand of a resource
          while still acknowledging the pressure from the size of
          population on the demand of a resource
- fix: correct methods in AmountResourceGood, VehicleGood,
          EquipmentGood, and RobotGood

Related to https://github.com/mars-sim/mars-sim/issues/1610 and
https://github.com/mars-sim/mars-sim/issues/1447