Fix intermarket average demand and average Good value
2025-07-23

- new: add two maps, namely, interMarketBook and settlementMarketBook
       in UnitManager
- new: add getInterMarketDemand() and getInterMarketGoodValue() in
       UnitManager
- change: revise unitUpdate() in TradeTableModel to display the average
          demand and average good value from all settlements
- change: fine tune initial params in AmountResourceGood
- change: Stabilize demand score for large settlement
- change: show decimal places of how much initial resources are stored
          in BuildingPanelStorage
- change: adjust regolith probability and modifier

Note: this commit uses UnitManager's thread to compute and to store
the average demand and average good value from all markets/settlements.
However, it won't show the correct score/values until the start of
sol 2. For now, in sol 1, it will only show the score/value of 0.1 on
all goods.

Related to https://github.com/mars-sim/mars-sim/issues/1610 and
https://github.com/mars-sim/mars-sim/issues/1447