Lower structural heat loss due to promixity of adjacent building
2025-07-25

- new: add proximityBarrierScore, HALLWAY_PROXIMITY, BUILDING_PROXIMITY
- change: calculate proximityBarrierScore for each building for
          reducing uValueAreaWall and uValueAreaCrackLength
          which in turns reduces structural heat loss in Heating

Note 1: this proximity effect helps buildings in the Sector Base fire
and foremost since a lot of buildings are next to each other.

Note 2: adjacentBuildings map is not accurate. If there's no connector
between two buildings, then it's NOT on the map. Will need to revisit
this in future to create an alternative proximity building map.

Related to https://github.com/mars-sim/mars-sim/issues/1625.